Package and clarify internal VLC support
Ensure python-vlc is installed and explicitly bundled by both Windows build modes, clarify that Internal VLC uses the locally installed VLC engine, and improve the unavailable-player guidance. Share normalized title matching with the internal player's playlist filter and bump the application version to V2.01.08.
